Begun re-factoring, going to take much more rabbit holes to canning it for today and going to implment new features before re-factoring
