+++ Commands/exampleCommand.py
@@ -1,0 +1,8 @@
+from discord.ext import commands
+
+@commands.command()
+async def CommandNameHERE(ctx):
+	COMMAND CODE HERE
+
+def setup(bot):
+	bot.add_command(CommandNameHERE)
